feat(shade-app): todos without re-render
